Move IndexerBase to abstract namespace
